# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Feature Flag Service",
    "description": "A minimal but serious feature flag platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/rules": {
      "post": {
        "tags": [
          "rules"
        ],
        "summary": "Create Rule",
        "operationId": "create_rule_api_v1_rules_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "rules"
        ],
        "summary": "List Rules",
        "operationId": "list_rules_api_v1_rules_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "flag_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Flag Id"
            }
          },
          {
            "name": "env",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Env"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RuleResponse"
                  },
                  "title": "Response List Rules Api V1 Rules Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Predicate": {
        "properties": {
          "attribute": {
            "type": "string",
            "minLength": 1,
            "title": "Attribute"
          },
          "operator": {
            "type": "string",
            "pattern": "^(exists|equals|not_equals|contains|in_list|gt|gte|lt|lte)$",
            "title": "Operator"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              },
              {
                "items": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "integer"
                    },
                    {
                      "type": "number"
                    }
                  ]
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "attribute",
          "operator"
        ],
        "title": "Predicate"
      },
      "RuleCreate": {
        "properties": {
          "flag_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Flag Id"
          },
          "flag_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Flag Key"
          },
          "environment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Environment Id"
          },
          "env_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Env Key"
          },
          "priority": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Priority",
            "default": 0
          },
          "conditions": {
            "items": {
              "$ref": "#/components/schemas/Predicate"
            },
            "type": "array",
            "title": "Conditions"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "variant": {
            "type": "string",
            "title": "Variant",
            "default": "on"
          }
        },
        "type": "object",
        "title": "RuleCreate"
      },
      "RuleResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "flag_id": {
            "type": "string",
            "title": "Flag Id"
          },
          "environment_id": {
            "type": "string",
            "title": "Environment Id"
          },
          "priority": {
            "type": "integer",
            "title": "Priority"
          },
          "conditions": {
            "items": {
              "$ref": "#/components/schemas/Predicate"
            },
            "type": "array",
            "title": "Conditions"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "variant": {
            "type": "string",
            "title": "Variant"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "flag_id",
          "environment_id",
          "priority",
          "conditions",
          "enabled",
          "variant",
          "created_at"
        ],
        "title": "RuleResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}